Be more willing to show Canadian dances for Canadian postal codes
The Canadian lookup system for TryContra is a bit of a hack: we only understand the first letter of the postal code and map it to a point, and so will think a dance is 'far' from the postal code the user has supplied even if it is actually in the same city. Fixing this properly is hard and involves understanding the Canadian system better.  But we can be more willing to show Canadian dances to users with Canadian postal codes, which fixes the cases I've tried (including V8X 4A9).